STORY/1833 - Client should log back in automatically after losing connection

Made a logged out flag for the connection.

Related work items: #1833

+++ src/assets/hooks/userHooks.js
@@ -117,8 +117,10 @@
     });
 
     hooks.Register(events.Socket.Closed, (e) => {
-        sessionStorage.clear();
-        store.commit("replaceEntireState", services.DefaultState());
+        if (connection.LoggedOut) {
+            sessionStorage.clear();
+            store.commit("replaceEntireState", services.DefaultState());
+        }
     });
